+++ e2e-tests/setup_test.go
@@ -2,7 +2,9 @@ package e2e_test
 
 import (
 	"bytes"
+	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/interchain-security/x/ccv/utils"
 
 	transfertypes "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
@@ -114,6 +116,98 @@ func (suite *ProviderTestSuite) SetupTest() {
 	suite.transferPath.EndpointB.ChannelConfig.PortID = transfertypes.PortID
 	suite.transferPath.EndpointA.ChannelConfig.Version = transfertypes.Version
 	suite.transferPath.EndpointB.ChannelConfig.Version = transfertypes.Version
+}
+
+// TODO: Can this be consolidated with ProviderTestSuite above?
+type KeeperTestSuite struct {
+	suite.Suite
+	coordinator *ibctesting.Coordinator
+
+	// testing chains
+	providerChain *ibctesting.TestChain
+	consumerChain *ibctesting.TestChain
+	path          *ibctesting.Path
+	ctx           sdk.Context
+}
+
+func (suite *KeeperTestSuite) SetupTest() {
+	suite.coordinator, suite.providerChain, suite.consumerChain = simapp.NewProviderConsumerCoordinator(suite.T())
+
+	// valsets must match
+	providerValUpdates := tmtypes.TM2PB.ValidatorUpdates(suite.providerChain.Vals)
+	consumerValUpdates := tmtypes.TM2PB.ValidatorUpdates(suite.consumerChain.Vals)
+	suite.Require().True(len(providerValUpdates) == len(consumerValUpdates), "initial valset not matching")
+	for i := 0; i < len(providerValUpdates); i++ {
+		addr1 := utils.GetChangePubKeyAddress(providerValUpdates[i])
+		addr2 := utils.GetChangePubKeyAddress(consumerValUpdates[i])
+		suite.Require().True(bytes.Equal(addr1, addr2), "validator mismatch")
+	}
+
+	// move both chains to the next block
+	suite.providerChain.NextBlock()
+	suite.consumerChain.NextBlock()
+
+	// create consumer client on provider chain and set as consumer client for consumer chainID in provider keeper.
+	err := suite.providerChain.App.(*appProvider.App).ProviderKeeper.CreateConsumerClient(
+		suite.providerChain.GetContext(),
+		suite.consumerChain.ChainID,
+		suite.consumerChain.LastHeader.GetHeight().(clienttypes.Height),
+		false,
+	)
+	suite.Require().NoError(err)
+	// move provider to next block to commit the state
+	suite.providerChain.NextBlock()
+
+	// initialize the consumer chain with the genesis state stored on the provider
+	consumerGenesis, found := suite.providerChain.App.(*appProvider.App).ProviderKeeper.GetConsumerGenesis(
+		suite.providerChain.GetContext(),
+		suite.consumerChain.ChainID,
+	)
+	suite.Require().True(found, "consumer genesis not found")
+	suite.consumerChain.App.(*appConsumer.App).ConsumerKeeper.InitGenesis(suite.consumerChain.GetContext(), &consumerGenesis)
+
+	// create path for the CCV channel
+	suite.path = ibctesting.NewPath(suite.consumerChain, suite.providerChain)
+
+	// update CCV path with correct info
+	// - set provider endpoint's clientID
+	consumerClient, found := suite.providerChain.App.(*appProvider.App).ProviderKeeper.GetConsumerClientId(
+		suite.providerChain.GetContext(),
+		suite.consumerChain.ChainID,
+	)
+	suite.Require().True(found, "consumer client not found")
+	suite.path.EndpointB.ClientID = consumerClient
+	// - set consumer endpoint's clientID
+	providerClient, found := suite.consumerChain.App.(*appConsumer.App).ConsumerKeeper.GetProviderClientID(suite.consumerChain.GetContext())
+	suite.Require().True(found, "provider client not found")
+	suite.path.EndpointA.ClientID = providerClient
+	// - client config
+	providerUnbondingPeriod := suite.providerChain.App.(*appProvider.App).GetStakingKeeper().UnbondingTime(suite.providerChain.GetContext())
+	suite.path.EndpointB.ClientConfig.(*ibctesting.TendermintConfig).UnbondingPeriod = providerUnbondingPeriod
+	suite.path.EndpointB.ClientConfig.(*ibctesting.TendermintConfig).TrustingPeriod = providerUnbondingPeriod / utils.TrustingPeriodFraction
+	consumerUnbondingPeriod := utils.ComputeConsumerUnbondingPeriod(providerUnbondingPeriod)
+	suite.path.EndpointA.ClientConfig.(*ibctesting.TendermintConfig).UnbondingPeriod = consumerUnbondingPeriod
+	suite.path.EndpointA.ClientConfig.(*ibctesting.TendermintConfig).TrustingPeriod = consumerUnbondingPeriod / utils.TrustingPeriodFraction
+	// - channel config
+	suite.path.EndpointA.ChannelConfig.PortID = consumertypes.PortID
+	suite.path.EndpointB.ChannelConfig.PortID = providertypes.PortID
+	suite.path.EndpointA.ChannelConfig.Version = types.Version
+	suite.path.EndpointB.ChannelConfig.Version = types.Version
+	suite.path.EndpointA.ChannelConfig.Order = channeltypes.ORDERED
+	suite.path.EndpointB.ChannelConfig.Order = channeltypes.ORDERED
+
+	// set chains sender account number
+	// TODO: to be fixed in #151
+	err = suite.path.EndpointB.Chain.SenderAccount.SetAccountNumber(6)
+	suite.Require().NoError(err)
+	err = suite.path.EndpointA.Chain.SenderAccount.SetAccountNumber(1)
+	suite.Require().NoError(err)
+
+	suite.ctx = suite.providerChain.GetContext()
+}
+
+func TestKeeperTestSuite(t *testing.T) {
+	suite.Run(t, new(KeeperTestSuite))
 }
 
 func (suite *ProviderTestSuite) SetupCCVChannel() {
