+++ e2e-tests/slashing_test.go
@@ -1,6 +1,7 @@
 package e2e_test
 
 import (
+	"fmt"
 	"time"
 
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
@@ -20,6 +21,7 @@ import (
 	"github.com/cosmos/interchain-security/x/ccv/types"
 
 	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/crypto/ed25519"
 )
 
 // TestSendDowntimePacket tests consumer initiated slashing
@@ -268,4 +270,228 @@ func (s *ProviderTestSuite) TestSlashPacketAcknowldgement() {
 
 	err = consumerKeeper.OnAcknowledgementPacket(s.consumerCtx(), packet, channeltypes.NewErrorAcknowledgement("another error"))
 	s.Require().Error(err)
+}
+
+// TestHandleSlashPacketDoubleSigning tests the handling of a double-signing related slash packet, with e2e tests
+func (suite *KeeperTestSuite) TestHandleSlashPacketDoubleSigning() {
+	providerKeeper := suite.providerChain.App.(*appProvider.App).ProviderKeeper
+	providerSlashingKeeper := suite.providerChain.App.(*appProvider.App).SlashingKeeper
+	providerStakingKeeper := suite.providerChain.App.(*appProvider.App).StakingKeeper
+
+	tmVal := suite.providerChain.Vals.Validators[0]
+	consAddr := sdk.ConsAddress(tmVal.Address)
+
+	// check that validator bonded status
+	validator, found := providerStakingKeeper.GetValidatorByConsAddr(suite.ctx, consAddr)
+	suite.Require().True(found)
+	suite.Require().Equal(stakingtypes.Bonded, validator.GetStatus())
+
+	// set init VSC id for chain0
+	providerKeeper.SetInitChainHeight(suite.ctx, suite.consumerChain.ChainID, uint64(suite.ctx.BlockHeight()))
+
+	// set validator signing-info
+	providerSlashingKeeper.SetValidatorSigningInfo(
+		suite.ctx,
+		consAddr,
+		slashingtypes.ValidatorSigningInfo{Address: consAddr.String()},
+	)
+
+	_, err := providerKeeper.HandleSlashPacket(suite.ctx, suite.consumerChain.ChainID,
+		ccv.NewSlashPacketData(
+			abci.Validator{Address: tmVal.Address, Power: 0},
+			uint64(0),
+			stakingtypes.DoubleSign,
+		),
+	)
+	suite.NoError(err)
+
+	// verify that validator is jailed in the staking and slashing mdodules' states
+	suite.Require().True(providerStakingKeeper.IsValidatorJailed(suite.ctx, consAddr))
+
+	signingInfo, _ := providerSlashingKeeper.GetValidatorSigningInfo(suite.ctx, consAddr)
+	suite.Require().True(signingInfo.JailedUntil.Equal(evidencetypes.DoubleSignJailEndTime))
+	suite.Require().True(signingInfo.Tombstoned)
+}
+
+// TestHandleSlashPacketErrors tests errors for the HandleSlashPacket method in an e2e testing setting
+func (suite *KeeperTestSuite) TestHandleSlashPacketErrors() {
+	providerStakingKeeper := suite.providerChain.App.(*appProvider.App).StakingKeeper
+	ProviderKeeper := suite.providerChain.App.(*appProvider.App).ProviderKeeper
+	providerSlashingKeeper := suite.providerChain.App.(*appProvider.App).SlashingKeeper
+	consumerChainID := suite.consumerChain.ChainID
+
+	// sync contexts block height
+	suite.ctx = suite.providerChain.GetContext()
+
+	// expect an error if initial block height isn't set for consumer chain
+	_, err := ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, ccv.SlashPacketData{})
+	suite.Require().Error(err, "slash validator with invalid infraction height")
+
+	// save VSC ID
+	vID := ProviderKeeper.GetValidatorSetUpdateId(suite.ctx)
+
+	// set faulty block height for current VSC ID
+	ProviderKeeper.SetValsetUpdateBlockHeight(suite.ctx, vID, 0)
+
+	// expect an error if block height mapping VSC ID is zero
+	_, err = ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, ccv.SlashPacketData{ValsetUpdateId: vID})
+	suite.Require().Error(err, "slash with height mapping to zero")
+
+	// construct slashing packet with non existing validator
+	slashingPkt := ccv.NewSlashPacketData(
+		abci.Validator{Address: ed25519.GenPrivKey().PubKey().Address(),
+			Power: int64(0)}, uint64(0), stakingtypes.Downtime,
+	)
+
+	// Set initial block height for consumer chain
+	ProviderKeeper.SetInitChainHeight(suite.ctx, consumerChainID, uint64(suite.ctx.BlockHeight()))
+
+	// expect the slash to not succeed if validator doesn't exist
+	success, err := ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, slashingPkt)
+	suite.Require().NoError(err, "slashing an unknown validator should not result in error")
+	suite.Require().False(success, "did slash unknown validator")
+
+	// jail an existing validator
+	val := suite.providerChain.Vals.Validators[0]
+	consAddr := sdk.ConsAddress(val.Address)
+	providerStakingKeeper.Jail(suite.ctx, consAddr)
+	// commit block to set VSC ID
+	suite.coordinator.CommitBlock(suite.providerChain)
+	// Update suite.ctx bc CommitBlock updates only providerChain's current header block height
+	suite.ctx = suite.providerChain.GetContext()
+	suite.Require().NotZero(ProviderKeeper.GetValsetUpdateBlockHeight(suite.ctx, vID))
+
+	// create validator signing info
+	valInfo := slashingtypes.NewValidatorSigningInfo(sdk.ConsAddress(val.Address), suite.ctx.BlockHeight(),
+		suite.ctx.BlockHeight()-1, time.Time{}.UTC(), false, int64(0))
+	providerSlashingKeeper.SetValidatorSigningInfo(suite.ctx, sdk.ConsAddress(val.Address), valInfo)
+
+	// update validator address and VSC ID
+	slashingPkt.Validator.Address = val.Address
+	slashingPkt.ValsetUpdateId = vID
+
+	// expect to slash and jail validator
+	_, err = ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, slashingPkt)
+	suite.Require().NoError(err, "did slash jail validator")
+
+	// expect error when infraction type in unspecified
+	tmAddr := suite.providerChain.Vals.Validators[1].Address
+	slashingPkt.Validator.Address = tmAddr
+	slashingPkt.Infraction = stakingtypes.InfractionEmpty
+
+	valInfo.Address = sdk.ConsAddress(tmAddr).String()
+	providerSlashingKeeper.SetValidatorSigningInfo(suite.ctx, sdk.ConsAddress(tmAddr), valInfo)
+
+	_, err = ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, slashingPkt)
+	suite.Require().EqualError(err, fmt.Sprintf("invalid infraction type: %v", stakingtypes.InfractionEmpty))
+
+	// expect to slash jail validator
+	slashingPkt.Infraction = stakingtypes.DoubleSign
+	_, err = ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, slashingPkt)
+	suite.Require().NoError(err)
+
+	// expect the slash to not succeed when validator is tombstoned
+	success, _ = ProviderKeeper.HandleSlashPacket(suite.ctx, consumerChainID, slashingPkt)
+	suite.Require().False(success)
+}
+
+// TestHandleSlashPacketDistribution tests the slashing of an undelegation balance
+// by varying the slash packet VSC ID mapping to infraction heights
+// lesser, equal or greater than the undelegation entry creation height
+func (suite *KeeperTestSuite) TestHandleSlashPacketDistribution() {
+	providerStakingKeeper := suite.providerChain.App.(*appProvider.App).StakingKeeper
+	providerKeeper := suite.providerChain.App.(*appProvider.App).ProviderKeeper
+
+	// choose a validator
+	tmValidator := suite.providerChain.Vals.Validators[0]
+	valAddr, err := sdk.ValAddressFromHex(tmValidator.Address.String())
+	suite.Require().NoError(err)
+
+	validator, found := providerStakingKeeper.GetValidator(suite.providerChain.GetContext(), valAddr)
+	suite.Require().True(found)
+
+	// unbonding operations parameters
+	delAddr := suite.providerChain.SenderAccount.GetAddress()
+	bondAmt := sdk.NewInt(1000000)
+
+	// new delegator shares used
+	testShares := sdk.Dec{}
+
+	// setup the test with a delegation, a no-op and an undelegation
+	setupOperations := []struct {
+		fn func(suite *KeeperTestSuite) error
+	}{
+		{
+			func(suite *KeeperTestSuite) error {
+				testShares, err = providerStakingKeeper.Delegate(suite.providerChain.GetContext(), delAddr, bondAmt, stakingtypes.Unbonded, stakingtypes.Validator(validator), true)
+				return err
+			},
+		}, {
+			func(suite *KeeperTestSuite) error {
+				return nil
+			},
+		}, {
+			// undelegate a quarter of the new shares created
+			func(suite *KeeperTestSuite) error {
+				_, err = providerStakingKeeper.Undelegate(suite.providerChain.GetContext(), delAddr, valAddr, testShares.QuoInt64(4))
+				return err
+			},
+		},
+	}
+
+	// execute the setup operations, distributed uniformly in three blocks.
+	// For each of them, save their current VSC Id value which map correspond respectively
+	// to the block heights lesser, equal and greater than the undelegation creation height.
+	vscIDs := make([]uint64, 0, 3)
+	for _, so := range setupOperations {
+		err := so.fn(suite)
+		suite.Require().NoError(err)
+
+		vscIDs = append(vscIDs, providerKeeper.GetValidatorSetUpdateId(suite.providerChain.GetContext()))
+		suite.providerChain.NextBlock()
+	}
+
+	// create validator signing info to test slashing
+	suite.providerChain.App.(*appProvider.App).SlashingKeeper.SetValidatorSigningInfo(
+		suite.providerChain.GetContext(),
+		sdk.ConsAddress(tmValidator.Address),
+		slashingtypes.ValidatorSigningInfo{Address: tmValidator.Address.String()},
+	)
+
+	// the test cases verify that only the unbonding tokens get slashed for the VSC ids
+	// mapping to the block heights before and during the undelegation otherwise not.
+	testCases := []struct {
+		expSlash bool
+		vscID    uint64
+	}{
+		{expSlash: true, vscID: vscIDs[0]},
+		{expSlash: true, vscID: vscIDs[1]},
+		{expSlash: false, vscID: vscIDs[2]},
+	}
+
+	// save unbonding balance before slashing tests
+	ubd, found := providerStakingKeeper.GetUnbondingDelegation(suite.providerChain.GetContext(), delAddr, valAddr)
+	suite.Require().True(found)
+	ubdBalance := ubd.Entries[0].Balance
+
+	for _, tc := range testCases {
+		slashPacket := ccv.NewSlashPacketData(
+			abci.Validator{Address: tmValidator.Address, Power: tmValidator.VotingPower},
+			tc.vscID,
+			stakingtypes.Downtime,
+		)
+
+		// slash
+		_, err := providerKeeper.HandleSlashPacket(suite.providerChain.GetContext(), suite.consumerChain.ChainID, slashPacket)
+		suite.Require().NoError(err)
+
+		ubd, found := providerStakingKeeper.GetUnbondingDelegation(suite.providerChain.GetContext(), delAddr, valAddr)
+		suite.Require().True(found)
+
+		isUbdSlashed := ubdBalance.GT(ubd.Entries[0].Balance)
+		suite.Require().True(tc.expSlash == isUbdSlashed)
+
+		// update balance
+		ubdBalance = ubd.Entries[0].Balance
+	}
 }
